requirements.txt
absl-py==0.9.0
appdirs==1.4.4
arrow==0.15.8
asgiref==3.2.3
astunparse==1.6.3
beautifulsoup4==4.8.2
bs4==0.0.1
cachetools==4.1.1
cairocffi==1.2.0
CairoSVG==2.5.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
configparser==4.0.2
crayons==0.3.0
cryptography==2.9
cssselect2==0.4.1
defusedxml==0.6.0
Django==3.0.3
django-allauth==0.41.0
django-appconf==1.0.4
django-compressor==2.4.1
django-cors-headers==3.2.1
django-environ==0.4.5
django-rest-auth==0.9.5
django-utils-six==2.0
djangorestframework==3.11.0
docopt==0.6.2
et-xmlfile==1.0.1
Faker==4.0.2
gast==0.3.3
google-auth==1.20.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.30.0
h5py==2.10.0
html5lib==1.1
idna==2.9
inform==1.21.0
jdcal==1.4.1
Keras-Preprocessing==1.1.2
Markdown==3.2.2
numpy==1.18.5
oauthlib==3.1.0
openpyxl==3.0.3
opt-einsum==3.3.0
pandas==1.1.0
Pillow==7.0.0
protobuf==3.12.4
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycrypto==2.6.1
PyJWT==1.7.1
Pyphen==0.10.0
python-dateutil==2.8.1
python-memcached==1.59
python3-openid==3.1.0
pytz==2019.3
quantiphy==2.12.0
rcssmin==1.0.6
reportlab==3.5.67
requests==2.23.0
requests-oauthlib==1.3.0
rjsmin==1.1.0
rsa==4.6
scipy==1.4.1
selenium==3.141.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
soupsieve==2.0
sqlparse==0.3.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow-addons==0.11.2
tensorflow-estimator==2.3.0
termcolor==1.1.0
text-unidecode==1.3
tflearn==0.3.2
tinycss2==1.1.0
tvm==1.0.0
typeguard==2.9.1
urllib3==1.25.8
WeasyPrint==52.5
webdriver-manager==2.3.0
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==4.0
wrapt==1.12.1
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
